Rework Then extensions, add docs to ObjectExtensions
Remove explicit Try to implicit, add more then methods, decompose classes to Then and ThenOnFail
