+++ tests/EffectiveResult.Tests/Extensions/ThenExtensionsTests.cs
@@ -1,11 +1,80 @@
-﻿using EffectiveResult.Abstractions;
-using EffectiveResult.Extensions;
+﻿using EffectiveResult.Extensions;
 using EffectiveResult.TestsCommon.Helpers;
 
 namespace EffectiveResult.Tests.Extensions;
 
 public class ThenExtensionsTests
 {
+    [Fact]
+    public void ThenExtension_WhenInvokeOnSuccessResult_ShouldBeInvoked()
+    {
+        // Arrange
+        var result = Result.Ok();
+        var flag = false;
+        Action action = () => flag = true;
+
+        // Act
+        var thenResult = result.Then(action);
+
+        // Assert
+        thenResult.Should().Be(result);
+        flag.Should().BeTrue();
+    }
+
+    [Fact]
+    public void ThenExtension_WhenInvokeOnFailedResult_ShouldNotBeInvoked()
+    {
+        // Arrange
+        var error = new Error("Bad");
+        var result = Result.Fail(error);
+
+        var flag = false;
+        Action action = () => flag = true;
+
+        // Act
+        var thenResult = result.Then(action);
+
+        // Assert
+        thenResult.Should().Be(result);
+        flag.Should().BeFalse();
+    }
+
+    [Fact]
+    public void ThenExtension_WhenInvokeOnSuccessTypedResult_ShouldBeInvoked()
+    {
+        // Arrange
+        var value = "Hello there!";
+        var result = Result.Ok(value);
+
+        string? expected = null;
+        Action<string> action = x => expected = x;
+
+        // Act
+        var thenResult = result.Then(action);
+
+        // Assert
+        thenResult.ShouldBeSuccess();
+        expected.Should().Be(value);
+    }
+
+    [Fact]
+    public void ThenExtension_WhenInvokeOnFailedTypedResult_ShouldNotBeInvoked()
+    {
+        // Arrange
+        var error = new Error("Deadlock");
+        var result = Result.Fail<string>(error);
+
+        string? expected = null;
+        Action<string> action = x => expected = x;
+
+        // Act
+        var thenResult = result.Then(action);
+
+        // Assert
+        thenResult.ShouldBeFailed();
+        expected.Should().BeNull();
+    }
+
     [Fact]
     public void ThenExtension_WhenInvokeOnSuccessResultAndFunc_ShouldBeInvokedAndReturnTypedResult()
     {
@@ -329,147 +398,5 @@ public class ThenExtensionsTests
         thenResultFailed.ShouldBeFailed(error);
 
         expectedValues.Should().BeEmpty();
-    }
-
-    [Fact]
-    public void ThenOnFailExtension_WhenInvokeOnSuccessResultWithFuncFactory_ShouldNotBeInvoked()
-    {
-        // Arrange
-        var value = "Good time need good result";
-        var result = Result.Ok(value);
-
-        var internalValue = "Hello there!";
-        var isInvoked = false;
-        var action = () =>
-        {
-            isInvoked = true;
-            return internalValue;
-        };
-
-        var internalResult = Result.Ok(internalValue);
-        var isInvokedResult = false;
-        var actionResult = () =>
-        {
-            isInvokedResult = true;
-            return internalResult;
-        };
-
-        // Act
-        var firstResult = result.ThenOnFail(action);
-        var secondResult = result.ThenOnFail(actionResult);
-
-        // Assert
-        firstResult.Should().Be(result);
-        secondResult.Should().Be(result);
-        isInvoked.Should().BeFalse();
-        isInvokedResult.Should().BeFalse();
-    }
-
-    [Fact]
-    public void ThenOnFailExtension_WhenInvokeOnFailedResultWithFuncFactory_ShouldBeInvokedAndReturnNewResult()
-    {
-        // Arrange
-        var error = new Error("Deadlocker");
-        var result = Result.Fail<string>(error);
-
-        var internalValue = "Hello there!";
-        var isInvoked = false;
-        var action = () =>
-        {
-            isInvoked = true;
-            return internalValue;
-        };
-
-        var internalResult = Result.Ok(internalValue);
-        var isInvokedResult = false;
-        var actionResult = () =>
-        {
-            isInvokedResult = true;
-            return internalResult;
-        };
-
-        // Act
-        var firstResult = result.ThenOnFail(action);
-        var secondResult = result.ThenOnFail(actionResult);
-
-        // Assert
-        firstResult.Should().NotBe(result);
-        secondResult.Should().NotBe(result);
-        isInvoked.Should().BeTrue();
-        isInvokedResult.Should().BeTrue();
-
-        firstResult.ShouldBeSuccessAndReferenceEqualsValue(internalValue);
-        secondResult.ShouldBeSuccessAndReferenceEqualsValue(internalValue);
-    }
-
-    [Fact]
-    public void ThenOnFailExtension_WhenInvokeOnSuccessResultWithFuncByErrors_ShouldNotBeInvoked()
-    {
-        // Arrange
-        var value = "Good time need good result";
-        var result = Result.Ok(value);
-
-        var internalValue = "Hello there!";
-        var isInvoked = false;
-        var action = (IReadOnlyCollection<IError> errors) =>
-        {
-            isInvoked = true;
-            return internalValue;
-        };
-
-        var internalResult = Result.Ok(internalValue);
-        var isInvokedResult = false;
-        var actionResult = (IReadOnlyCollection<IError> errors) =>
-        {
-            isInvokedResult = true;
-            return internalResult;
-        };
-
-        // Act
-        var firstResult = result.ThenOnFail(action);
-        var secondResult = result.ThenOnFail(actionResult);
-
-        // Assert
-        firstResult.Should().Be(result);
-        secondResult.Should().Be(result);
-        isInvoked.Should().BeFalse();
-        isInvokedResult.Should().BeFalse();
-    }
-
-    [Fact]
-    public void ThenOnFailExtension_WhenInvokeOnFailedResultWithFuncByErrors_ShouldBeInvokedAndReturnNewResult()
-    {
-        // Arrange
-        var error = new Error("Deadlocker");
-        var result = Result.Fail<string>(error);
-
-        var internalValue = "Hello there!";
-        var isInvoked = false;
-        var action = (IReadOnlyCollection<IError> errors) =>
-        {
-            isInvoked = true;
-            return internalValue;
-        };
-
-        var internalResult = Result.Ok(internalValue);
-        var isInvokedResult = false;
-        var actionResult = (IReadOnlyCollection<IError> errors) =>
-        {
-            isInvokedResult = true;
-            return internalResult;
-        };
-
-        // Act
-        var firstResult = result.ThenOnFail(action);
-        var secondResult = result.ThenOnFail(actionResult);
-
-        // Assert
-        firstResult.Should().NotBe(result);
-        secondResult.Should().NotBe(result);
-        isInvoked.Should().BeTrue();
-        isInvokedResult.Should().BeTrue();
-
-        firstResult.ShouldBeSuccessAndReferenceEqualsValue(internalValue);
-        secondResult.ShouldBeSuccessAndReferenceEqualsValue(internalValue);
     }
 }
